src/flask_app/requirements.txt
Flask==1.1.2
setuptools==44.0.0
wheel==0.33.6
pytest==5.3.2
coverage==5.0.2
mysql-connector-python==8.0.18
